requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
pandas
wandb
ultralytics
cv2